+++ wumgr/StartupElevationPolicy.cs
@@ -1,0 +1,10 @@
+namespace wumgr
+{
+    internal static class StartupElevationPolicy
+    {
+        public static bool ShouldAttemptStartupElevation(bool isAdministrator, bool isDebugging, bool skipUacConfigured)
+        {
+            return !isAdministrator && !isDebugging && skipUacConfigured;
+        }
+    }
+}